requirements.txt
apache-airflow==2.9.1
pandas==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==1.4.49
requests==2.31.0
ezodf==0.3.2
odfpy==1.4.1
pyexcel-ods3==0.6.1
pyarrow>=15.0.0
fastparquet>=2024.2.0

